Movde classes from src/index.js to lib/<ClassName> and require instead
